requirements.txt
streamlit==1.29.0
openai==1.7.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
requests==2.31.0
